+++ src/store/socket.js
@@ -176,6 +176,10 @@ class LiveSession {
         if (!this._isSpectator) return;
         this._store.commit("toggleNight", params);
         break;
+      case "isReturnToTown":
+        if (!this._isSpectator) return;
+        this._store.commit("toggleReturnToTown");
+        break;
       case "isVoteHistoryAllowed":
         if (!this._isSpectator) return;
         this._store.commit("session/setVoteHistoryAllowed", params);
@@ -209,6 +213,12 @@ class LiveSession {
       case "pronouns":
         this._updatePlayerPronouns(params);
         break;
+      case "emote":
+        this._updateEmote(params);
+        break;
+      case "emoteType":
+        this._updateEmoteType(params);
+        break;
     }
   }
 
@@ -221,9 +231,7 @@ class LiveSession {
     if (!this._store.state.session.playerId) {
       this._store.commit(
         "session/setPlayerId",
-        Math.random()
-          .toString(36)
-          .substr(2)
+        Math.random().toString(36).substr(2)
       );
     }
     this._pings = {};
@@ -265,6 +273,8 @@ class LiveSession {
       isDead: player.isDead,
       isVoteless: player.isVoteless,
       pronouns: player.pronouns,
+      isEmote: player.isEmote,
+      emoteType: player.emoteType,
       ...(player.role && player.role.team === "traveler"
         ? { roleId: player.role.id }
         : {}),
@@ -281,6 +291,7 @@ class LiveSession {
       this._sendDirect(playerId, "gs", {
         gamestate: this._gamestate,
         isNight: grimoire.isNight,
+        isReturnToTown: grimoire.isReturnToTown,
         isVoteHistoryAllowed: session.isVoteHistoryAllowed,
         isVoteWatchingAllowed: session.isVoteWatchingAllowed,
         nomination: session.nomination,
@@ -305,6 +316,7 @@ class LiveSession {
       gamestate,
       isLightweight,
       isNight,
+      isReturnToTown,
       isVoteHistoryAllowed,
       isVoteWatchingAllowed,
       nomination,
@@ -331,7 +343,15 @@ class LiveSession {
       const player = players[x];
       const { roleId } = state;
       // update relevant properties
-      ["name", "id", "isDead", "isVoteless", "pronouns"].forEach((property) => {
+      [
+        "name",
+        "id",
+        "isDead",
+        "isVoteless",
+        "pronouns",
+        "isEmote",
+        "emoteType",
+      ].forEach((property) => {
         const value = state[property];
         if (player[property] !== value) {
           this._store.commit("players/update", { player, property, value });
@@ -359,6 +379,7 @@ class LiveSession {
     });
     if (!isLightweight) {
       this._store.commit("toggleNight", !!isNight);
+      this._store.commit("toggleReturnToTown", !!isReturnToTown);
       this._store.commit("session/setVoteHistoryAllowed", isVoteHistoryAllowed);
       this._store.commit(
         "session/setVoteWatchingAllowed",
@@ -532,6 +553,65 @@ class LiveSession {
     this._send("pronouns", [index, value]);
   }
 
+  /**
+   * Update isEmote based on incoming data.
+   * @param player
+   * @param value
+   * @param isFromSockets
+   */
+  sendIsEmote({ player, value, isFromSockets }) {
+    if (isFromSockets || this._store.state.session.playerId !== player.id)
+      return;
+    const index = this._store.state.players.players.indexOf(player);
+    this._send("emote", [index, value]);
+  }
+
+  /**
+   * Update EmoteType based on incoming data.
+   * @param player
+   * @param value
+   * @param isFromSockets
+   */
+  sendEmoteType({ player, value, isFromSockets }) {
+    if (isFromSockets || this._store.state.session.playerId !== player.id)
+      return;
+    const index = this._store.state.players.players.indexOf(player);
+    this._send("emoteType", [index, value]);
+  }
+
+  /**
+   * Update a hand raised state based on incoming data.
+   * @param index
+   * @param value
+   * @private
+   */
+  _updateEmote([index, value]) {
+    const player = this._store.state.players.players[index];
+
+    this._store.commit("players/update", {
+      player,
+      property: "isEmote",
+      value,
+      isFromSockets: true,
+    });
+  }
+
+  /**
+   * Update a emote type state based on incoming data.
+   * @param index
+   * @param value
+   * @private
+   */
+  _updateEmoteType([index, value]) {
+    const player = this._store.state.players.players[index];
+
+    this._store.commit("players/update", {
+      player,
+      property: "emoteType",
+      value,
+      isFromSockets: true,
+    });
+  }
   /**
    * Update a pronouns based on incoming data.
    * @param index
@@ -711,6 +791,11 @@ class LiveSession {
   setIsNight() {
     if (this._isSpectator) return;
     this._send("isNight", this._store.state.grimoire.isNight);
+  }
+
+  setReturnToTown() {
+    if (this._isSpectator) return;
+    this._send("isReturnToTown", this._store.state.grimoire.isReturnToTown);
   }
 
   /**
@@ -908,6 +993,8 @@ export default (store) => {
       case "toggleNight":
         session.setIsNight();
         break;
+      case "toggleReturnToTown":
+        session.setReturnToTown();
       case "setEdition":
         session.sendEdition();
         break;
@@ -934,6 +1021,10 @@ export default (store) => {
       case "players/update":
         if (payload.property === "pronouns") {
           session.sendPlayerPronouns(payload);
+        } else if (payload.property === "isEmote") {
+          session.sendIsEmote(payload);
+        } else if (payload.property === "emoteType") {
+          session.sendEmoteType(payload);
         } else {
           session.sendPlayer(payload);
         }